removed not sender and note factory
